seperate external network tests
