+++ .github/workflows/test-external-network.yml
@@ -1,0 +1,19 @@
+name: "tests / external network"
+
+# on:
+#   schedule:
+#     - cron: "0 6 * * *"
+#   workflow_dispatch:
+on:
+  pull_request:
+    branches: [ '**' ]
+
+jobs:
+  real-network:
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v6.0.2
+      - name: run real network tests
+        run: bash ./scripts/test-external-network
