+++ scripts/test-external-network
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+cd `dirname $0`/..
+
+docker compose -f docker-compose.test.yaml build test
+docker compose -f docker-compose.test.yaml up -d test-redis test-grafana
+docker compose -f docker-compose.test.yaml run test python -m unittest discover test.external_network
